// src/types/starship.ts
export type Starship = {
  name: string;
  model: string;
  manufacturer: string;
  cost_in_credits: string | null;
  length: string;
  max_atmosphering_speed: string | null;
  crew: string | null;
  passengers: string | null;
  cargo_capacity: string | null;
  consumables: string | null;
  hyperdrive_rating: string;
  MGLT: string | null;
  starship_class: string;
  pilots: string[];
  films: string[];
  created: string;
  edited: string;
  url: string;
}; 